(REF)(TEST) separate different surface scalar tests
